% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_proposal.R
\name{bin_proposal}
\alias{bin_proposal}
\title{Propose according to binomial distribution}
\usage{
bin_proposal(n, x, p)
}
\arguments{
\item{n}{number of particles.}

\item{x}{number of successes.}

\item{p}{success probability.}
}
\value{
list containing: number of trials, corresponding log-likelihood.
}
\description{
Generates proposals for unknown number of trials n given x successes and success probability p.
}
\examples{
bin_proposal(n=100, x=10, p=0.1)
}
