% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_nodes_cpp}
\alias{dist_nodes_cpp}
\title{Distances from Leaves to Root of Phylogenetic Tree}
\usage{
dist_nodes_cpp(ptree_nodes, ptree_lengths, root_node)
}
\arguments{
\item{ptree_nodes}{matrix of nodes and edge lengths.}

\item{root_node}{label of root node.}
}
\value{
distances from leaves to root
}
\description{
Computes the distance from the leaves to the root of a phylogenetic tree.
}
\examples{
dist_nodes_cpp <- function(ptree_nodes, root_node)

}
