% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distToRoot.R
\name{distToRoot}
\alias{distToRoot}
\title{Distance to root}
\usage{
distToRoot(phy)
}
\arguments{
\item{phy}{object of class phylo.}
}
\value{
vector containing distance from leaf to root.
}
\description{
For each leave on the tree, calculate the distance from the leaf to the root.
}
\examples{
distToRoot(ptree)
}
