% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtnbinom.R
\name{dtnbinom}
\alias{dtnbinom}
\title{Density function for the truncated negative binomial distribution}
\usage{
dtnbinom(x, size, mu, a, b, log = TRUE)
}
\arguments{
\item{x}{vector of quantities.}

\item{size}{dispersion parameter.}

\item{mu}{mean.}

\item{a, b}{lower and upper truncation points (a <= x <= b).}

\item{log}{logical; if TRUE, probabilities are given as log(p).}
}
\value{
density.
}
\description{
Density function for the truncated negative binomial distribution.
}
\examples{
dtnbinom(x = 1, size = 10, mu = 2, a = 1, b = 100)
}
