% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_fns.R
\name{epi_fns}
\alias{epi_fns}
\alias{constant}
\alias{change_pt}
\alias{linear_dec}
\alias{linear_inc}
\title{A collection of functions for the birth rate in used to generate epidemics}
\usage{
constant(t, stop_time, a = 0.3)

change_pt(t, stop_time, cp = 0.5, a = 0.5, b = 0.1)

linear_dec(t, stop_time, a = 0.5, b = 0.1)

linear_inc(t, stop_time, a = 0.1, b = 0.5)
}
\arguments{
\item{t}{time}

\item{stop_time}{length of the epidemic}

\item{a}{start rate}

\item{b}{end rate}
}
\value{
function
}
\description{
Epidemic functions
}
