% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic.R
\name{epidemic}
\alias{epidemic}
\title{Simulate epidemics}
\usage{
epidemic(birth_rate, death_rate, stop_time)
}
\arguments{
\item{birth_rate}{birth rate of the epidemic.}

\item{death_rate}{death rate of the epidemic.}

\item{stop_time}{number of days to run the epidemic simulation for.}
}
\value{
data frame with: Index, Parent Index, Infection Status, Infection Time, Removal Time
}
\description{
Simulates a birth-death epidemic with constant birth and death rates and 1 infected on day 0.
}
\examples{
epidemic(birth_rate = 0.2, death_rate = 0.1, stop_time = 50)
}
