% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nopt_noadapt.R
\name{find_nopt_noadapt}
\alias{find_nopt_noadapt}
\title{Find optimal number of particles}
\usage{
find_nopt_noadapt(
  sigma0,
  proportion_obs0,
  x0 = 1,
  death_rate,
  ptree,
  day = 0,
  noisy_prevalence,
  resampling_scheme = "systematic",
  print = F
)
}
\arguments{
\item{sigma0}{initial value of the linear gaussian standard deviation.}

\item{proportion_obs0}{initial value of the proportion of cases observed.}

\item{x0}{prevalence on day 0.}

\item{death_rate}{death rate of the epidemic.}

\item{ptree}{object of class phylo.}

\item{day}{number of days in the past the most recent leaf was sampled.}

\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{resampling_scheme}{"multinomial" or "systematic".}

\item{print}{logical; if TRUE, prints percentage of the way through the chain.}
}
\value{
optimal number of particles
}
\description{
Find optimal number of particles
}
\examples{
find_nopt_noadapt(sigma0 = 0.1, proportion_obs0 = 0.5, death_rate = 0.1, ptree = sample_tree, day = 1, noisy_prevalence = noisy_prev, print = T)
}
