% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_llik.R
\name{genetic_llik}
\alias{genetic_llik}
\title{Genetic Log-Likelihood}
\usage{
genetic_llik(birth_rate, ptree, prevalence, stop_time, log = T)
}
\arguments{
\item{birth_rate}{birth rate of the epidemic.}

\item{ptree}{object of class phylo.}

\item{prevalence}{data frame of prevalences by day.}

\item{stop_time}{number of days the epidemic simulation was run for.}

\item{log}{logical; if TRUE, log-likelihood is given.}
}
\value{
log-likelihood.
}
\description{
Probability of the phylogenetic tree given the birth rate and the prevalence.
}
\examples{
genetic_llik(birth_rate = 0.2, ptree = sample_tree, prevalence = prev, stop_time = 50)
}
