% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggiNEXTPD}
\alias{ggiNEXTPD}
\title{Plots the outcome of iNEXTPD using \code{ggplot2} package.}
\usage{
ggiNEXTPD(outcome, plot.type = 1:3, datatype = "abundance")
}
\arguments{
\item{outcome}{the outcome of the function \code{iNEXTPD}}

\item{plot.type}{a positive integer vector specifying types of curves. Three types of plots: sample-size-based rarefaction and extrapolation curve (\code{plot.type = 1});
sample coverage curve (\code{plot.type = 2});coverage-based rarefaction and extrapolation curve (\code{plot.type = 3}). Default is \code{c(1,2,3)} \cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}). Default is "abundance". \cr}
}
\value{
plots of the estimated curves based on \code{ggplot2} package. Different choice of \code{plot.type} will yied different types of plot:
\itemize{
 \item{Sample-size-based R/E curve (\code{plot.type = 1}): the curve of estimates as a function of sample size.} \cr
 \item{Sample completeness curve (\code{plot.type = 2}): the curve of sample coverage with respect to sample size.} \cr
 \item{Coverage-based R/E curve (\code{plot.type = 3}): the curve of the estimates as a function of sample coverage.} \cr
 }
}
\description{
\code{ggiNEXTPD} plot the outcome of iNEXTPD using ggplot2 package.
}
\examples{
\donttest{
# Type (1) abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- iNEXTPD(data = data, tree = tree,datatype = "abundance",q = c(0,1,2))
ggiNEXTPD(out)
# Type (2) incidence data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
t_ <- data.inc$t
out <- iNEXTPD(data = data, t_ = t_,datatype = "incidence_raw", tree = tree,q = c(0,1,2))
ggiNEXTPD(out,datatype = "incidence_raw")
}
}
