% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggtqplot}
\alias{ggtqplot}
\title{Plot the outcome of \code{PhdObs} or \code{PhdAsy} based on \code{ggplot2} package.}
\usage{
ggtqplot(outcome, profile = "q")
}
\arguments{
\item{outcome}{the outcome of the function \code{PhdObs} or \code{PhdAsy}.}

\item{proflie}{specifying the type of profile: \code{profile = 'q'} for q profile and \code{profile = 'time'} for time profile.}
}
\value{
plot of the PD/D empirical or estimated curves based on \code{ggplot2} package.
}
\description{
\code{ggtqplot}: plot the outcome of \code{PhdObs} or \code{PhdAsy} using \code{ggplot2} package.
}
\examples{
\donttest{
# Type (1) q profile
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- PhdObs(data = data,datatype = "abundance",tree = tree,q = seq(0, 2, by = 0.25))
ggtqplot(out,profile = 'q')
# Type (2) time profile
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- PhdObs(data = data,datatype = "abundance",tree = tree,q = c(0,1,2),
reftime = seq(0.1,325,length.out = 40))
ggtqplot(out,profile = 'time')
}
}
