% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{inextPD_yhc}
\alias{inextPD_yhc}
\title{Interpolation and extrapolation of phylogenetic diversity}
\usage{
inextPD_yhc(data, tree, datatype = "abundance", t_, q = c(0, 1, 2),
  endpoint = NULL, knots = 40, size = NULL, plot.type = 1:3,
  conf = 0.95, nboot = 50, reftime = NULL)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr Type (1) abundance data: When there are N assemblages, the
observed species abundances should be arranged as a species (in rows) by assemblage (in columns) matrix. The first row
(including N entries) lists the assemblage labels or site names for the N assemblages.\cr Type (2) incidence data:
The data input format for incidence data must be raw detection/non-detection data. That is, data for each community/assemblage
consist of a species-by-sampling-unit matrix. Users must first merge multiple-community data by species identity to obtain a pooled
list of species; then the rows of the input data refer to this pooled list. \cr}

\item{tree}{a phylogenetic tree of class phylo among the observed species in the pooled assemblage.\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}), default is "abundance". \cr}

\item{t_}{a nammed vector of nonnegative integers specifying the sampling units in each community. Ignored if \code{datatype = "abundance"}.\cr}

\item{q}{a nonnegative integer value or integer vector specifying the diversity order of Hill numbers, default is 0. \cr}

\item{endpoint}{an interger specifying the endpoint for rarefaction and extrapolation range. If \code{NULL}, \code{endpoint} = double of the maximum
reference sample size. It will be ignored if \code{size} is given. \cr}

\item{knots}{an integer specifying the number of knot between 1 and the \code{endpoint}, default is 40.\cr}

\item{size}{a vector of nonnegative integers specifying the sample sizes for which diversity estimates will be calculated. If \code{NULL}, the diversity estimates will
be calculated for those sample sizes determined by the specified/default \code{endpoint} and \code{knot}. \cr}

\item{plot.type}{an integer or an integer vector specifying types of curves. Three types of plots: sample-size-based rarefaction and extrapolation curve (\code{plot.type = 1}); coverage-based rarefaction
and extrapolation curve (\code{plot.type = 2}); sample completeness curve (\code{plot.type = 3}), default is \code{plot.type = 1}. \cr}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95. \cr}

\item{nboot}{an integer specifying the number of bootstrap replications. Enter 0 to skip bootstrap; in this case,
the computation of s.e. and confidence intervals will be skipped, default is 50.}

\item{reftime}{a positive number specifying the reference time.}
}
\value{
a list of four objects: \cr\cr
\code{$summary} summary of data. \cr\cr
\code{$refernce.point} root of phylogeny tree for the species in the pooled assemblage. \cr\cr
\code{$inext} table of various estimates for interpolated or extrapolated samples and their confidence intervals. \cr\cr
\code{$figure} showing chosen \code{plot.type} of sampling curves. \cr\cr
}
\description{
\code{inextPD_yhc}: Interpolation and extrapolation of phylogenetic diversities based on a framework of Hill numbers with order q.
}
\examples{
\donttest{
# Type (1) abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- inextPD_yhc(data = data, tree = tree,datatype = "abundance",q = c(0,1,2), plot.type = 1:3)
# Type (2) incidence data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
t_ <- data.inc$t
out <- inextPD_yhc(data = data, tree = tree,datatype = "incidence_raw",t_ = t_,q = c(0,1,2), plot.type = 1:3)
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. Philosophical Transactions of the Royal Society B., 365, 3599-3609.\cr\cr
Chao, A., Chiu, C.-H., Hsieh, T. C., Davis, T., Nipperess, D., and Faith, D. (2015) Rarefaction and extrapolation of phylogenetic diversity. Methods in Ecology and Evolution, 6, 380-388.\cr\cr
}
