% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_lse.R
\name{pairwise_lse}
\alias{pairwise_lse}
\title{Pairwise logSumExp}
\usage{
pairwise_lse(la, lb)
}
\arguments{
\item{la}{log(a).}

\item{lb}{log(b).}
}
\value{
log(a + b)
}
\description{
Implements logSumExp trick on pairs.
}
\examples{
pairwise_lse(la = logq + epi_prior, lb = log(1-q) + epi_data)
}
