% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{PDInfo}
\alias{PDInfo}
\title{Summarize phylogenetic data infomation.}
\usage{
PDInfo(data, t_, datatype = "abundance", tree, reftime = NULL)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr
See \code{\link{iNEXTPD}} for data details.}

\item{t_}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the sampling units in each assemblage. Ignored if \code{datatype = "abundance"}.\cr
Ignored if \code{datatype = "abundance"}.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}). Default is "abundance". \cr}

\item{tree}{a phylo object describing the Newick phylogeny tree for all observed species in the pooled assemblage. \cr}

\item{reftime}{a positive value or sequence specifying the reference time for tree. If \code{NULL}, \code{reftime} = the tree depth of pooled assemblage. Default is \code{NULL}.}
}
\value{
a tibble of phylogenetic data statistics, including reference sample size \code{n}/ number of sampling units \code{T},
number of observed species \code{S.obs}, observed branch length (Faith's PD) \code{PD.obs}, the first two species frequency counts
(\code{f1*/Q1*}, \code{f2*/Q2*}) and their branch length sums, respectively (\code{g1/R1}, \code{g2/R2}) at specified/default reference time \code{reftime}. \cr\cr
}
\description{
\code{PDInfo}: summarize phylogenetic data statistics for specified/default reference time \code{reftime}.
}
\examples{
\donttest{
# Type (1) abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- PDInfo(data = data,datatype = "abundance", tree = tree)
# Type (2) incidence data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
t_ <- data.inc$t
out <- PDInfo(data = data, t_ = t_, datatype = "incidence_raw", tree = tree)
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. Philosophical Transactions of the Royal Society B., 365, 3599-3609. \cr\cr
Chao, A., Chiu, C.-H., Hsieh, T. C., Davis, T., Nipperess, D., and Faith, D. (2015) Rarefaction and extrapolation of phylogenetic diversity. Methods in Ecology and Evolution, 6, 380-388.\cr\cr
Hsieh, T. C. and Chao, A. (2017). Rarefaction and extrapolation: making fair comparison of abundance-sensitive phylogenetic diversity among multiple assemblages. Systematic Biology 66, 100-111.
}
