% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{PhdAsy_yhc}
\alias{PhdAsy_yhc}
\title{Asymptotic phylogenetic diversity profile for each community}
\usage{
PhdAsy_yhc(data, tree, datatype = "abundance", t, q = seq(0, 2, by =
  0.25), conf = 0.95, nboot = 50, reftime = NULL)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr
See \code{\link{inextPD_yhc}} for data details.}

\item{tree}{a phylogenetic tree of class phylo among the observed species in the pooled assemblage. \cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}), default is "abundance". \cr}

\item{t}{a vector of nonnegative integers specifying the sampling units in each community.
Ignored if \code{datatype = "abundance"}.}

\item{q}{a nonnegative vector specifying the diversity order of Hill numbers. For phylogenetic diversity profiles,
\code{q} need to be a vector of length more than one, default is seq(0, 2, by = 0.25) \cr}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95. \cr}

\item{nboot}{an integer specifying the number of bootstrap replications. Enter 0 to skip bootstrap; in this case,
the computation of s.e. and confidence intervals will be skipped, default is 50.}

\item{reftime}{a positive number specifying the reference time.}
}
\value{
a list of four objects: \cr\cr
\code{$summary} summary of data. \cr\cr
\code{$refernce.point} root of the species in the pooled assemblage. \cr\cr
\code{$asy} a table showing all numerical values of diversity estimates and their s.e and confidence intervals. \cr\cr
\code{$figure} phylogenetic diversity profiles. \cr\cr
}
\description{
\code{PhdAsy}: An estimated asymptotic phylogenetic diversity profile for each community based on a framework of Hill numbers with order q and statistical method proposed in Chao et al. (2015) and Hsieh and Chao (2017).\cr
}
\examples{
\donttest{
# Type (1) abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- PhdAsy_yhc(data = data, tree = tree, q = seq(0, 2, by = 0.25))
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. Philosophical Transactions of the Royal Society B., 365, 3599-3609. \cr\cr
}
