% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{PhdObs_yhc}
\alias{PhdObs_yhc}
\title{Observed q profile and time profile of phylogenetic diversity or phylogenetic Hill numbers for each community}
\usage{
PhdObs_yhc(data, tree, datatype = "abundance", t_, type = "PD",
  profile = "q", q = seq(0, 2, by = 0.25), tprofile_times = NULL,
  nboot = 50, conf = 0.95, knots = 0, reftime = NULL)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr
See \code{\link{inextPD_yhc}} for data details.}

\item{tree}{a phylogenetic tree in the Newick format among the observed species in the pooled assemblage. \cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}), default is "abundance". \cr}

\item{t_}{a named vector of nonnegative integers specifying the sampling units in each community. If the names are absent, each community will be
automatically assigned siteX as community names.
Ignored if \code{datatype = "abundance"}.}

\item{type}{two types for calculating phylogenetic profiles in Chiu, C.-H., Jost, L. and Chao(2014): Phylogenetic Diversity(\code{type = "PD"}) and
Phylogenetic Hill numbers(\code{type = "D"}).}

\item{profile}{two kinds of profiles to show phylogenetic profiles based on chosen \code{type}: q profile(\code{profile = "q"}) and time profile(\code{profile = "time"}). \cr}

\item{q}{a nonnegative vector specifying the diversity order of Hill numbers. For phylogenetic diversity profiles,
\code{q} need to be a vector of length more than one, default is seq(0, 2, by = 0.25) \cr}

\item{tprofile_times}{a nonnegative vector specifying the reference time points for which phylogenetic profiles will be computed.\cr}

\item{nboot}{an integer specifying the number of bootstrap replications. Enter 0 to skip bootstrap; in this case,
the computation of s.e. and confidence intervals will be skipped, default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.}

\item{knots}{an integer specifying the number of points of \code{time} used to calculate AUC (Area Under Curve). Use 0 to skip calculation, default is 0\cr}

\item{reftime}{a positive number specifying the reference time.}
}
\value{
Based on chosen \code{knots}, there are two kinds of output:\cr\cr
\itemize{
 \item{For \code{knots} an integer, a list of fourth objects:} \cr
\code{$summary} summary of data. \cr\cr
\code{$fortime_table} or \code{$forq_table} a table of observed phylogenetic diversity for each \code{time} or order \code{q} based on \code{type} and \code{profile}. \cr\cr
\code{$fortime_figure} or \code{$forq_figure} showing the curves for \code{$fortime_table} or \code{$forq_table} \cr\cr
\code{$AUC_table} table of Area Under Curve (AUC) of time profile for each site from 0.01 to root. \cr\cr
 \item{For \code{knots = NULL}, a list of three objects:} \cr
\code{$summary} for summarizing data information. \cr\cr
\code{$fortime_table} or \code{$forq_table} a table of observed estimates for each \code{time} or order \code{q} based on \code{type} and \code{profile}. \cr\cr
\code{$fortime_figure} or \code{$forq_figure} showing the curves for \code{$fortime_table} or \code{$forq_table} \cr\cr
}
}
\description{
\code{PhdObs}: Observed phylogenetic diversity or phylogenetic Hill numbers profile for each community based on a framework of Hill numbers with order q and reference time.\cr
}
\examples{
\donttest{
# Type (1) abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- PhdObs_yhc(data = data, tree = tree, datatype = "abundance", type = "PD", profile = "q", q = seq(0, 2, by = 0.25))
# Type (2) incidence data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
t_ <- data.inc$t
out <- PhdObs_yhc(data = data, tree = tree, datatype = "incidence_raw", type = "PD", profile = "q", q = seq(0, 2, by = 0.25))
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. Philosophical Transactions of the Royal Society B., 365, 3599-3609. \cr\cr
}
