% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_tree.R
\name{phylo_tree}
\alias{phylo_tree}
\title{Phylogenetic Tree}
\usage{
phylo_tree(epidemic, stop_time)
}
\arguments{
\item{epidemic}{output from epidemic().}

\item{stop_time}{number of days the epidemic simulation was run for.}
}
\value{
object of class phylo.
}
\description{
Represents the epidemic as a phylogenetic tree.
}
\examples{
phylo_tree(epidemic = epi, stop_time = 50)
}
