% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_noepi.R
\name{pm_noepi}
\alias{pm_noepi}
\title{Pseudo-Marginal MCMC with No Epidemic Data}
\usage{
pm_noepi(
  iter,
  birth_rate_0,
  max_birth_rate = 100,
  prevalence_0,
  death_rate,
  ptree,
  n_particles
)
}
\arguments{
\item{iter}{number of iterations to run the algorithm for.}

\item{birth_rate_0}{initial value of the birth rate.}

\item{max_birth_rate}{maximum value for the birth rate. 100 by default.}

\item{prevalence_0}{data frame of initial values for the prevalence per day.}

\item{death_rate}{death rate of the epidemic.}

\item{ptree}{object of class phylo.}

\item{n_particles}{number of particles used in the importance sampling.}
}
\value{
list containing: birth rate, acceptance rate, run time in seconds
}
\description{
Implements a pseudo-marginal MCMC when no epidemic data has been observed.
}
\examples{
pm_noepi(iter = 100000, birth_rate_0 = 0.1, prevalence_0 = data.frame("day"=0:50, "prev"=rep(1,51)). death_rate = 0.1, ptree = sample_tree, n_particles = 100)
}
