% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_noepi_parallel.R
\name{pm_noepi_parallel}
\alias{pm_noepi_parallel}
\title{Pseudo-Marginal MCMC with No Epidemic Data - Parallelised}
\usage{
pm_noepi_parallel(
  iter,
  birth_rate_0,
  max_birth_rate = 100,
  prevalence_0,
  death_rate,
  ptree,
  n_particles,
  n_cores,
  print = F
)
}
\arguments{
\item{iter}{number of iterations to run the algorithm for.}

\item{birth_rate_0}{initial value of the birth rate.}

\item{max_birth_rate}{maximum value for the birth rate. 100 by default.}

\item{prevalence_0}{data frame of initial values for the prevalence per day.}

\item{death_rate}{death rate of the epidemic.}

\item{ptree}{object of class phylo.}

\item{n_particles}{number of particles used in the importance sampling.}

\item{n_cores}{number of cores to use in parallelisation.}

\item{print}{logical; if TRUE, prints percentage of the way through the chain.}
}
\value{
list containing: birth rate, acceptance rate, run time in seconds
}
\description{
Parallelised version of pm_noepi().
}
\examples{
pm_noepi_parallel(iter = 100000, birth_rate_0 = 0.1, prevalence_0 = data.frame("day"=0:50, "prev"=rep(1,51)). death_rate = 0.1, ptree = sample_tree, n_particles = 100, n_cores = 8)
}
