% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_noisy.R
\name{pm_noisy}
\alias{pm_noisy}
\title{Pseudo-Marginal MCMC with Partial Epidemic Data}
\usage{
pm_noisy(
  iter,
  birth_rate_0,
  max_birth_rate = 100,
  prevalence_0,
  death_rate,
  ptree,
  noisy_prevalence,
  proportion_obs,
  n_particles,
  print = F
)
}
\arguments{
\item{iter}{number of iterations to run the algorithm for.}

\item{birth_rate_0}{initial value of the birth rate.}

\item{max_birth_rate}{maximum value for the birth rate. 100 by default.}

\item{prevalence_0}{data frame of initial values for the prevalence per day.}

\item{death_rate}{death rate of the epidemic.}

\item{ptree}{object of class phylo.}

\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{proportion_obs}{proportion of cases observed.}

\item{n_particles}{number of particles used in the importance sampling.}

\item{print}{logical; if TRUE, prints percentage of the way through the chain.}
}
\value{
list containing: birth rate, acceptance rate, run time in seconds
}
\description{
Implements a pseudo-marginal MCMC when the epidemic has been partially observed with known proportion.
}
\examples{
prev_0 <- data.frame("day"=0:50, "prev"=c(1, floor(noisy_prev[-1,2] / proportion_obs)))
prev_0[prev_0[,2] == 0, 2] <- 1
pm_noisy(iter = 100000, birth_rate_0 = 0.1, prevalence_0 = prev_0, death_rate = 0.1, ptree = sample_tree, noisy_prevalence = noisy_prev, proportion_obs = 0.2, n_particles = 100)
}
