% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Prevalence}
\usage{
prevalence(epidemic, stop_time)
}
\arguments{
\item{epidemic}{output from epidemic().}

\item{stop_time}{number of days the epidemic simulation was run for.}
}
\value{
data frame with 2 columns and N+1 rows. Column 1 is day and column 2 is number of cases.
}
\description{
Turns epidemic into a data frame of day and prevalence on that day.
}
\examples{
prevalence(epidemic = epi, stop_time = 50)
}
