% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_pois2.R
\name{propose_pois2}
\alias{propose_pois2}
\title{Propose Epidemic using Smoothed Poisson}
\usage{
propose_pois2(
  birth_rate,
  death_rate,
  noisy_prevalence,
  proportion_obs,
  ptree,
  ...
)
}
\arguments{
\item{birth_rate}{birth rate of the epidemic.}

\item{death_rate}{death rate of the epidemic.}

\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{proportion_obs}{proportion of cases observed.}

\item{ptree}{object of class phylo.}

\item{...}{allows for the use of lapply to run it multiple times.}
}
\value{
log-likelihood.
}
\description{
Simulates an epidemic by scaling up and smoothing the partially observed epidemic and sampling prevalence from a poisson distribution centred on this scaled and smoothed epidemic.
}
\examples{
propose_pois2(birth_rate = 0.2, death_rate = 0.1, noisy_prevalence = noisy_prev, proportion_obs = 0.2, ptree = sample_tree)
}
