% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtskellam.R
\name{rtskellam}
\alias{rtskellam}
\title{Truncated Skellam distribution}
\usage{
rtskellam(n, old_x, birth_rate, death_rate)
}
\arguments{
\item{n}{number of observations.}

\item{old_x}{old prevalence.}

\item{birth_rate}{birth rate of the epidemic.}

\item{death_rate}{death rate of the epidemic.}
}
\value{
log-likelihood.
}
\description{
Random generation truncated at 0 for the Skellam distribution
}
\examples{
rtskellam(n = 100, old_x = x_resample, birth_rate = b_sample, death_rate = 0.1)
}
