% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_het.R
\name{sample_het}
\alias{sample_het}
\title{Sample from Infected}
\usage{
sample_het(ptree, pi)
}
\arguments{
\item{ptree}{object of class phylo.}

\item{pi}{probability of sampling a leaf.}

\item{day}{vector of days to sample from.}
}
\value{
object of class phylo; the tree keeping only the sampled leaves.
}
\description{
Samples and keeps leaves with probability pi.
}
\examples{
sample_het(ptree = full_tree, pi = 0.1)
}
