% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_inf.R
\name{sample_inf}
\alias{sample_inf}
\title{Sample from Infected}
\usage{
sample_inf(ptree, pi)
}
\arguments{
\item{ptree}{object of class phylo.}

\item{pi}{probability of sampling a leaf.}
}
\value{
object of class phylo; the tree keeping only the sampled leaves.
}
\description{
Samples and keeps those infected at the present day (i.e. the leaves) with probability pi.
}
\examples{
sample_inf(ptree = full_tree, pi = 0.1)
}
