% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_be.R
\name{sir_be}
\alias{sir_be}
\title{Sample Importance Resample - Adaptive}
\usage{
sir_be(
  n_particles,
  max_birth_rate = 10,
  death_rate,
  noisy_prevalence,
  proportion_obs,
  genetic_data,
  ess_threshold = n_particles/2,
  plot = F
)
}
\arguments{
\item{n_particles}{number of particles used in the sampling.}

\item{death_rate}{death rate of the epidemic.}

\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{proportion_obs}{proportion of cases observed.}

\item{genetic_data}{data frame of day, number of lineages and number of coalescences.}

\item{ess_threshold}{threshold of ESS below which triggers resampling.}

\item{plot}{logical; if TRUE, then plots a prevalence trajectory according to weight.}

\item{birth_rate_max}{maximum value of the birth rate on day 1 of the epidemic.}
}
\value{
log-likelihood
}
\description{
Implements an adaptive SIR algorithm and returns an approximated log-likelihood. Proposals are Skellam.
}
\examples{
sir_be(n_particles = 100, death_rate = 0.1, noisy_prevalence = noisy_prev, proportion_obs = 0.2, genetic_data = gen_data)
}
