% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellam_llik.R
\name{skellam_llik}
\alias{skellam_llik}
\title{Skellam Log-Likelihood}
\usage{
skellam_llik(birth_rate, death_rate, prevalence, log = T)
}
\arguments{
\item{birth_rate}{birth rate of the epidemic.}

\item{death_rate}{death rate of the epidemic.}

\item{prevalence}{data frame of prevalence per day.}

\item{log}{logical; if TRUE, log-likelihood is given.}
}
\value{
log-likelihood.
}
\description{
Approximates the probability of the prevalence given the birth and death rate.
}
\examples{
skellam_llik(birth_rate = 0.2, death_rate = 0.1, prevalence = prev)
}
