% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_nopt.R
\name{smc_nopt}
\alias{smc_nopt}
\title{PMMH with Unknown Number of Particles}
\usage{
smc_nopt(
  iter,
  max_time = Inf,
  max_birth_rate0,
  sigma0,
  proportion_obs0,
  death_rate,
  ptree,
  noisy_prevalence,
  n_particles = NULL,
  ess_threshold = n_particles/2,
  resampling_scheme = "systematic",
  backward_sim = TRUE,
  print = F
)
}
\arguments{
\item{iter}{number of iterations to run the algorithm for.}

\item{max_time}{maximum number of seconds to run the algorithm for.}

\item{max_birth_rate0}{initial value of the maximum birth rate.}

\item{sigma0}{initial value of the linear gaussian standard deviation.}

\item{proportion_obs0}{initial value of the proportion of cases observed.}

\item{death_rate}{death rate of the epidemic.}

\item{ptree}{object of class phylo.}

\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{n_particles}{number of particles used in the importance sampling.}

\item{ess_threshold}{threshold of ESS below which triggers resampling.}

\item{resampling_scheme}{"multinomial" or "systematic".}

\item{backward_sim}{logical; if TRUE, uses backward simulation.}

\item{print}{logical; if TRUE, prints percentage of the way through the chain.}
}
\value{
list containing: birth rate, prevalence, proportion observed, maximum birth rate, linear gaussian variance, acceptance rate, run time in seconds, smc log-likelihood and optimal number of particles
}
\description{
Implements an SMC-MCMC with adaptive resampling when the epidemic with birth rate varying over time has been partially observed with known proportion. MH adaptation of shape and scale. Unknown optimum number of particles.
}
\examples{
smc_nopt(iter = 100000, max_birth_rate0 = 1, sigma0 = 0.1, proportion_obs0 = 0.5, death_rate = 0.1, ptree = sample_tree, noisy_prevalence = noisy_prev, print = T)
}
