% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_prior.R
\name{smc_prior}
\alias{smc_prior}
\title{PMMH with Unknown Number of Particles}
\usage{
smc_prior(
  iter,
  max_time = Inf,
  sigma0,
  proportion_obs0,
  x0 = 1,
  death_rate,
  ptree,
  day = 0,
  genetic_data = NULL,
  noisy_prevalence,
  n_particles = NULL,
  ess_threshold = n_particles/2,
  max_n_particles = 10000,
  min_n_particles = 1000,
  resampling_scheme = "systematic",
  backward_sim = TRUE,
  print = F
)
}
\arguments{
\item{iter}{number of iterations to run the algorithm for.}

\item{max_time}{maximum number of seconds to run the algorithm for.}

\item{sigma0}{initial value of the linear gaussian standard deviation.}

\item{proportion_obs0}{initial value of the proportion of cases observed.}

\item{x0}{initial value of the prevalence on day 0.}

\item{death_rate}{death rate of the epidemic.}

\item{ptree}{object of class phylo.}

\item{day}{number of days in the past the most recent leaf was sampled.}

\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{n_particles}{number of particles used in the importance sampling.}

\item{ess_threshold}{threshold of ESS below which triggers resampling.}

\item{max_n_particles}{upper bound on the number of particles to use in the importance sampling.}

\item{min_n_particles}{lower bound on the number of particles to use in the importance sampling.}

\item{resampling_scheme}{"multinomial" or "systematic".}

\item{backward_sim}{logical; if TRUE, uses backward simulation.}

\item{print}{logical; if TRUE, prints percentage of the way through the chain.}
}
\value{
list containing: birth rate, prevalence, proportion observed, maximum birth rate, linear gaussian variance, acceptance rate, run time in seconds, smc log-likelihood and optimal number of particles
}
\description{
Implements an SMC-MCMC with adaptive resampling when the epidemic with birth rate varying over time has been partially observed with known proportion. MH adaptation of shape and scale of rho, sigma and X0. Unknown optimum number of particles.
}
\examples{
smc_prior(iter = 100000, sigma0 = 0.1, x0 = 1, proportion_obs0 = 0.5, death_rate = 0.1, ptree = sample_tree, day = 0, noisy_prevalence = noisy_prev, print = T)
}
