% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_skellam.R
\name{smc_skellam}
\alias{smc_skellam}
\title{Skellam Log-Likelihood}
\usage{
smc_skellam(new_x, old_x, birth_rate, death_rate, log = T)
}
\arguments{
\item{new_x}{new prevalence.}

\item{old_x}{old prevalence.}

\item{birth_rate}{birth rate of the epidemic.}

\item{death_rate}{death rate of the epidemic.}

\item{log}{logical; if TRUE, log-likelihood is given.}
}
\value{
log-likelihood.
}
\description{
Approximates the probability of a change in prevalence given the birth and death rate.
}
\examples{
smc_skellam(new_x = 10, old_x = 7, birth_rate = 0.2, death_rate = 0.1)
}
