% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_x0.R
\name{smc_x0}
\alias{smc_x0}
\title{PMMH with Unknown Number of Particles}
\usage{
smc_x0(
  iter,
  max_time = Inf,
  target_acceptance = 0.1,
  sigma0,
  proportion_obs0,
  x0 = 1,
  death_rate,
  ptree,
  day = 0,
  genetic_data = NULL,
  noisy_prevalence,
  sigma_mean = 0.1,
  pobs_prior = "uniform",
  pobs_min = 0,
  pobs_max = 1,
  pobs_alpha = 1,
  pobs_beta = 1,
  x0_prior = "uniform",
  x0_min = 1,
  x0_max = Inf,
  x0_mean = 10,
  x0_var = 100,
  n_particles = NULL,
  ess_threshold = n_particles/2,
  max_n_particles = 10000,
  min_n_particles = 1000,
  resampling_scheme = "systematic",
  backward_sim = TRUE,
  print = F
)
}
\arguments{
\item{iter}{number of iterations to run the algorithm for.}

\item{max_time}{maximum number of seconds to run the algorithm for.}

\item{target_acceptance}{target acceptance rate.}

\item{sigma0}{initial value of the linear gaussian standard deviation.}

\item{proportion_obs0}{initial value of the proportion of cases observed.}

\item{x0}{initial value of the prevalence on day 0.}

\item{death_rate}{death rate of the epidemic.}

\item{ptree}{object of class phylo.}

\item{day}{number of days in the past the most recent leaf was sampled.}

\item{genetic_data}{data frame of day, number of lineages and number of coalescences.}

\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{sigma_mean}{exponential prior mean of sigma.}

\item{pobs_prior}{"uniform" or "beta"; prior distribution on reporting probability.}

\item{pobs_min}{minimum for uniform reporting probability prior.}

\item{pobs_max}{maximum for uniform reporting probability prior.}

\item{pobs_alpha}{shape1 for beta reporting probability prior.}

\item{pobs_beta}{shape2 for beta reporting probability prior.}

\item{x0_prior}{"uniform" or "nbinom"; prior distribution on day 0 prevalence.}

\item{x0_min}{minimum for uniform day 0 prevalence prior.}

\item{x0_max}{maximum for uniform day 0 prevalence prior.}

\item{x0_mean}{mean for negative binomial day 0 prevalence prior.}

\item{x0_var}{variance for negative binomial day 0 prevalence prior.}

\item{n_particles}{number of particles used in the importance sampling.}

\item{ess_threshold}{threshold of ESS below which triggers resampling.}

\item{max_n_particles}{upper bound on the number of particles to use in the importance sampling.}

\item{min_n_particles}{lower bound on the number of particles to use in the importance sampling.}

\item{resampling_scheme}{"multinomial" or "systematic".}

\item{backward_sim}{logical; if TRUE, uses backward simulation.}

\item{print}{logical; if TRUE, prints percentage of the way through the chain.}
}
\value{
list containing: birth rate, prevalence, proportion observed, maximum birth rate, linear gaussian variance, acceptance rate, run time in seconds, smc log-likelihood and optimal number of particles
}
\description{
Implements an SMC-MCMC with adaptive resampling when the epidemic with birth rate varying over time has been partially observed with known proportion. MH adaptation of shape and scale of rho and sigma. Unknown optimum number of particles.
}
\examples{
smc_x0(iter = 100000, sigma0 = 0.1, x0 = 1, proportion_obs0 = 0.5, death_rate = 0.1, ptree = sample_tree, day = 0, noisy_prevalence = noisy_prev, print = T)
}
