% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Smoothing}
\usage{
smooth(noisy_prevalence, proportion_obs)
}
\arguments{
\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{proportion_obs}{proportion of cases observed.}
}
\value{
data frame of estimated prevalence per day.
}
\description{
Scales up and smooths the observed prevalence.
}
\examples{
smooth(noisy_prevalence = noisy_prev, proportion_obs = 0.2)
}
