% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{systematic_sample_cpp}
\alias{systematic_sample_cpp}
\title{Systematic Resampling}
\usage{
systematic_sample_cpp(n_particles, norm_weights)
}
\arguments{
\item{n_particles}{number of particles used in the sampling.}

\item{norm_weights}{normalised weights of the particles.}
}
\value{
resampled particles
}
\description{
Implements systematic resampling.
}
\examples{
systematic_sample_cpp <- function(n_particles = 1000, norm_weights = rep(1/1000, 1000))

}
