% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varying_bt.R
\name{varying_bt}
\alias{varying_bt}
\title{Adaptive SMC-MCMC with Partial Epidemic Data with Varying Birth Rates}
\usage{
varying_bt(
  iter,
  max_time = Inf,
  birth_rate0,
  max_birth_rate0,
  sigma0,
  proportion_obs0,
  death_rate,
  ptree,
  noisy_prevalence,
  n_particles,
  ess_threshold = n_particles/2,
  print = F
)
}
\arguments{
\item{iter}{number of iterations to run the algorithm for.}

\item{max_time}{maximum number of seconds to run the algorithm for.}

\item{birth_rate0}{initial value of the vector of birth rates.}

\item{max_birth_rate0}{initial value of the maximum birth rate.}

\item{sigma0}{initial value of the linear Gaussian standard deviation.}

\item{proportion_obs0}{initial value of the proportion of cases observed.}

\item{death_rate}{death rate of the epidemic.}

\item{ptree}{object of class phylo.}

\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{n_particles}{number of particles used in the importance sampling.}

\item{ess_threshold}{threshold of ESS below which triggers resampling.}

\item{print}{logical; if TRUE, prints percentage of the way through the chain.}
}
\value{
list containing: birth rate, proportion observed, maximum birth rate, linear gaussian variance, acceptance rate, run time in seconds and smc log-likelihood
}
\description{
Implements an SMC-MCMC with adaptive resampling when the epidemic with birth rate varying over time has been partially observed with known proportion.
}
\examples{
varying_bt(iter = 100000, birth_rate_0 = rep(0.1, 50), max_birth_rate0 = 1, sigma0 = 0.1, proportion_obs0 = 0.1, death_rate = 0.1, ptree = sample_tree, noisy_prevalence = noisy_prev, n_particles = 100)
}
