% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{estimatePD}
\alias{estimatePD}
\title{Compute phylogenetic diversity with a particular sample coverages.}
\usage{
estimatePD(data, tree, datatype = "abundance", t_, q = c(0, 1, 2),
  level = NULL, nboot = 50, conf = 0.95, reftime = NULL)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr
See \code{\link{iNEXTPD}} for data details.}

\item{tree}{a phylo object describing the Newick phylogeny tree for all observed species in the pooled assemblage. \cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}). Default is "abundance". \cr}

\item{t_}{needed only when datatype = "incidence_raw", a sequence of named nonnegative integers specifying the sampling units in each community. Ignored if \code{datatype = "abundance"}.\cr
Ignored if \code{datatype = "abundance"}.}

\item{q}{a nonnegative sequence specifying the diversity orders of PD. Default is \code{c(0,1,2)}. \cr}

\item{level}{a positive sequence < 1 specifying a particular value of sample coverage.
If \code{NULL},then \code{level} will be chosen as the minimum coverage of all sites after extrapolating each site to its double sample sizes. Default is \code{NULL}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications. Enter 0 to skip bootstrap;
in this case, the caculation of standard errors and confidence intervals will be skipped. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95. \cr}

\item{reftime}{a positive value specifying the reference time for tree. If \code{NULL}, \code{reftime} = the tree depth of pooled assemblage. Default is \code{NULL}.}
}
\value{
a \code{tibble} tabulating the PD estimates with each \code{q} for the user-specified sample coverage. In addition,
the corresponding sample sizes and sample coverages are provided. \cr\cr
}
\description{
\code{EstimatePD}: computes phylogenetic diversity(PD) with particular user-specified levels of sample coverage.
}
\examples{
\donttest{
# Type (1) abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- estimatePD(data = data, tree = tree,datatype = "abundance")
# Type (2) incidence data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
t_ <- data.inc$t
out <- estimatePD(data = data, tree = tree,datatype = "incidence_raw",t_ = t_)
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. Philosophical Transactions of the Royal Society B., 365, 3599-3609. \cr\cr
Chao, A., Chiu, C.-H., Hsieh, T. C., Davis, T., Nipperess, D., and Faith, D. (2015) Rarefaction and extrapolation of phylogenetic diversity. Methods in Ecology and Evolution, 6, 380-388.\cr\cr
Hsieh, T. C. and Chao, A. (2017). Rarefaction and extrapolation: making fair comparison of abundance-sensitive phylogenetic diversity among multiple assemblages. Systematic Biology 66, 100-111.
}
