% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{iNEXTPD}
\alias{iNEXTPD}
\title{Interpolation and extrapolation of phylogenetic diversity/phylogenetic Hill numbers}
\usage{
iNEXTPD(
  data,
  t_,
  datatype = "abundance",
  tree,
  q = c(0, 1, 2),
  reftime = NULL,
  type = "PD",
  endpoint = NULL,
  knots = 40,
  size = NULL,
  nboot = 50,
  conf = 0.95
)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr Type (1) abundance data: a S by N matrix/data.frame
where N is the number of assemblages. The element in i-th row and k-th is the abundance of species i in assemblage k. Please note
that the rownames of data must be the species names matching the species names in phylogeny tree and thus can not be empty.\cr
Type (2) incidence data: the sampling unit is quadrat or transect, the observed species was only recorded as presence(detection)/absence(non-detection)
data in each sampling unit. For single assemblage consisting of t_ sampling units, data is a S by t_ incidence matrix,
where S is the number of species. When there are N assemblages, users must first merge N data matrices by species identity to
obtain a large incidence matrix, where the rows of the matrix refer to all species presented in the pooled data. Likewise,
the rownames of data must be the species names matching the species names in phylogeny tree and thus can not be empty. \cr}

\item{t_}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the sampling units in each assemblage. Ignored if \code{datatype = "abundance"}.\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}), default is \code{"abundance"}. \cr}

\item{tree}{a phylo object describing the Newick phylogeny tree for all observed species in the pooled assemblage.\cr}

\item{q}{a nonnegative value or sequence specifying the diversity order. Default is \code{c(0,1,2)}.}

\item{reftime}{a positive value or sequence specifying the reference time for tree. If \code{NULL}, \code{reftime} = the tree depth of pooled assemblage. Default is \code{NULL}.}

\item{type}{desired diversity type: \code{type = "PD"} for phylogenetic diversity and \code{type = "D"} for phylogenetic Hill number. See Chao et al. (2010) for details. Default is \code{"PD"}.}

\item{endpoint}{an positive interger specifying the endpoint for rarefaction and
extrapolation range. If \code{NULL}, \code{endpoint} = double of the maximum reference sample size. It will be ignored if \code{size} is given. \cr}

\item{knots}{a positive integer specifying the number of knots between 1 and the \code{endpoint}. Default is 40.\cr}

\item{size}{a sequence of positive integers specifying the sample sizes for which \code{PD} or \code{D} estimates will be calculated. If \code{NULL}, then estimates will be
calculated for those sample sizes determined by the specified/default \code{endpoint} and \code{knots}. \cr}

\item{nboot}{a positive integer specifying the number of bootstrap replications. Enter 0 to skip bootstrap;
in this case, the caculation of standard errors and confidence intervals will be skipped. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95. \cr}
}
\value{
a list containg two tibbles:
 \code{$size_based} for sizd-based PD or D estimates and sample coverage for interpolated or extrapolated sample sizes along with their confidence intervals (if \code{nboot > 0}). \cr\cr
 \code{$coverage_based} for coverage-based one.
}
\description{
\code{iNEXTPD}: the seamless rarefaction and extrapolation sampling curves of phylogenetic diversity(PD) or phylogenetic Hill numbers(D).
See Chao et al. (2010, 2015) and Hsieh and Chao (2017) for pertinent background and methods.
}
\examples{
\donttest{
# Type (1) abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- iNEXTPD(data = data, tree = tree,datatype = "abundance",q = c(0,1,2))
# Type (2) incidence data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
t_ <- data.inc$t
out <- iNEXTPD(data = data, t_ = t_,datatype = "incidence_raw", tree = tree,q = c(0,1,2))
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. Philosophical Transactions of the Royal Society B., 365, 3599-3609.\cr\cr
Chao, A., Chiu, C.-H., Hsieh, T. C., Davis, T., Nipperess, D., and Faith, D. (2015) Rarefaction and extrapolation of phylogenetic diversity. Methods in Ecology and Evolution, 6, 380-388.\cr\cr
Hsieh, T. C. and Chao, A. (2017). Rarefaction and extrapolation: making fair comparison of abundance-sensitive phylogenetic diversity among multiple assemblages. Systematic Biology 66, 100-111.
}
