% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{PhdObs}
\alias{PhdObs}
\title{Observed phylogenetic diversity/phylogenetic Hill numbers}
\usage{
PhdObs(
  data,
  t_,
  datatype = "abundance",
  tree,
  q = seq(0, 2, by = 0.25),
  reftime = NULL,
  type = "PD",
  nboot = 50,
  conf = 0.95
)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr
See \code{\link{iNEXTPD}} for data details.}

\item{t_}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the numbers of sampling units in each assemblage. Ignored if \code{datatype = "abundance"}.\cr
If \code{names(t_) = NULL}, automatically assign siteX as assemblage names. Ignored if \code{datatype = "abundance"}.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}). Default is \code{"abundance"}. \cr}

\item{tree}{a phylo object describing the Newick phylogeny tree for all observed species in the pooled assemblage. \cr}

\item{q}{a nonnegative value or sequence specifying the diversity order. Default is \code{seq(0, 2, by = 0.25)}.}

\item{reftime}{a positive value or sequence specifying the reference time for tree. If \code{NULL}, \code{reftime} = the tree depth of pooled assemblage. Default is \code{NULL}.}

\item{type}{desired diversity type: \code{type = "PD"} for phylogenetic diversity and \code{type = "D"} for phylogenetic Hill number. See Chao et al. (2010) for details. Default is \code{"PD"}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications. Enter 0 to skip bootstrap;
in this case, the caculation of standard errors and confidence intervals will be skipped. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.}
}
\value{
a tibble of empirical(observed) phylogenetic diversity(\code{type = "PD"}) or
phylogenetic Hill number(\code{type = "D"}) with respect to each \code{reftime} and order \code{q}.\cr\cr
}
\description{
\code{PhdObs} computes emprircal phylogenetic diversity(PD) or phylogenetic Hill numbers(D) for specified/default order \code{q} and reference time \code{reftime};
see Chao et al. (2010) for details of PD and D.\cr
}
\examples{
\donttest{
# Type (1) abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- PhdObs(data = data,datatype = "abundance",tree = tree,
q = seq(0, 2, by = 0.25),reftime = c(162.5,325))
# Type (2) incidence data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
t_ <- data.inc$t
out <- PhdObs(data = data,t_ = t_, datatype = "incidence_raw",
tree = tree, q = seq(0, 2, by = 0.25))
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. Philosophical Transactions of the Royal Society B., 365, 3599-3609. \cr\cr
}
