% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_het.R
\name{sample_het}
\alias{sample_het}
\title{Sample from Infected}
\usage{
sample_het(ptree, day = 0, pi0, pi1)
}
\arguments{
\item{ptree}{object of class phylo.}

\item{day}{how many days in the past the most recent leaf is.}

\item{pi0}{probability of sampling a leaf at the present day.}

\item{pi1}{probability of sampling a leaf before the present day.}
}
\value{
object of class phylo; the tree keeping only the sampled leaves.
}
\description{
Samples and keeps leaves with probability pi.
}
\examples{
sample_het(ptree = full_tree, pi0 = 0, pi1 = 0.01)
}
