% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_be.R
\name{sir_be}
\alias{sir_be}
\title{Sample Importance Resample - Adaptive}
\usage{
sir_be(
  n_particles,
  max_birth_rate,
  sigma,
  death_rate,
  noisy_prevalence,
  proportion_obs,
  genetic_data,
  ess_threshold = n_particles/2,
  resampling_scheme = "multinomial",
  backward_sim = TRUE
)
}
\arguments{
\item{n_particles}{number of particles used in the sampling.}

\item{max_birth_rate}{maximum value of the birth rate on day 1 of the epidemic.}

\item{sigma}{standard deviation of the linear gaussian proposal for birth rates.}

\item{death_rate}{death rate of the epidemic.}

\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{proportion_obs}{proportion of cases observed.}

\item{genetic_data}{data frame of day, number of lineages and number of coalescences.}

\item{ess_threshold}{threshold of ESS below which triggers resampling.}

\item{resampling_scheme}{"multinomial" or "systematic".}

\item{backward_sim}{logical; if TRUE, uses backward simulation.}
}
\value{
log-likelihood
}
\description{
Implements an adaptive SIR algorithm and returns an approximated log-likelihood. Prevalence proposals are negative binomial and birth rate proposals are linear Gaussian.
}
\examples{
sir_be(n_particles = 100, max_birth_rate = 1, death_rate = 0.1, noisy_prevalence = noisy_prev, proportion_obs = 0.2, genetic_data = gen_data)
}
