% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varying_bt.R
\name{varying_bt}
\alias{varying_bt}
\title{Adaptive SMC-MCMC with Partial Epidemic Data with Varying Birth Rates}
\usage{
varying_bt(
  iter,
  birth_rate_0,
  max_birth_rate = 100,
  prevalence_0,
  death_rate,
  ptree,
  noisy_prevalence,
  proportion_obs,
  n_particles,
  ess_threshold = n_particles/2,
  print = F,
  plot = F
)
}
\arguments{
\item{iter}{number of iterations to run the algorithm for.}

\item{birth_rate_0}{initial value of the vector of birth rates.}

\item{max_birth_rate}{maximum value for the birth rate. 100 by default.}

\item{prevalence_0}{data frame of initial values for the prevalence per day.}

\item{death_rate}{death rate of the epidemic.}

\item{ptree}{object of class phylo.}

\item{noisy_prevalence}{data frame of observed prevalence per day.}

\item{proportion_obs}{proportion of cases observed.}

\item{n_particles}{number of particles used in the importance sampling.}

\item{ess_threshold}{threshold of ESS below which triggers resampling.}

\item{print}{logical; if TRUE, prints percentage of the way through the chain.}

\item{plot}{logical; if TRUE, then plots a trajectory according to weight.}
}
\value{
list containing: birth rate, acceptance rate, run time in seconds
}
\description{
Implements an SMC-MCMC with adaptive resampling when the epidemic with birth rate varying over time has been partially observed with known proportion.
}
\examples{
varying_bt(iter = 100000, birth_rate_0 = 0.1, prevalence_0 = prev_0, death_rate = 0.1, ptree = sample_tree, noisy_prevalence = noisy_prev, proportion_obs = 0.2, n_particles = 100)
}
